Add a route so the file can be actually downloaded by the browser
