+++ file_encrypt.services.yml
@@ -1,0 +1,9 @@
+services:
+  stream_wrapper.encrypt:
+    class: \Drupal\file_encrypt\EncryptStreamWrapper
+    tags:
+      - { name: stream_wrapper, scheme: encrypt }
+  path_processor.encrypt_files:
+    class: \Drupal\file_encrypt\PathProcessor\PathProcessorFiles
+    tags:
+      - { name: path_processor_inbound, priority: 200 }
